requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# other requirements
pandas
numpy
mlflow
pyyaml
argparse
psycopg2-binary
python-dotenv
requests
